[FEATURE] Basic models for elections and ballot boxes
This creates the basic model classes for elections and ballot boxes.
The election can have multiple election periods, which define when it is
active (and voting sessions will be accepted). Outside these periods,
no voting will be possible.

The ballot box model is rather dumb currently, it just contains the
basic properties and the different states it may have, but no methods
for transitioning the state. This should, once implemented, be further
secured with automated logging and functional tests.