[FEATURE] Basic version of EligibleVoters
The basic version of the EligibleVoter class holds the name of the voter
and an arbitrary number of attributes, called discriminators. The
discriminators are stored in a separate table and accessible with their
identifier. This currently requires a patched version of TYPO3 Flow,
therefore a functional test has been added for this feature.