[!!!][FEATURE] (CLI) identifier for elections
To be able to use elections in command-line scripts, we need a way to
address them on the command line. The regular way would be to use the
40-char UUID autogenerated by Flow/Doctrine for objects. This is however
really inconvenient and error-prone, so this commit introduces a new
identifier in the Election class, replacing
persistence_object_identifier.

The breaking change here is that this new identifier is used everywhere
and has to be set manually when creating the election. Due to its usage
as a foreign key constraint in DB tables, it cannot be changed after the
election has been created (well, or rather after related objects have
been created, but this is close enough to creation of the object
itself).

Using the included migration, it should be possible to seamlessly
migrate existing data, with the drawback that you still have to use the
unhandy 40-char UUIDs. This has however not been tested with real data
and is just an assumption.

For reasons of backwards compatibility, the election identifier can be
up to 40 characters, but that is not really recommended. Keep it
straight and simple…