[TASK] Remove relation from election to voters
An election possibly has a huge number of voters, in the tens of
thousands. Holding all these objects in the collection would mean a
huge memory usage and slow processing, while providing little benefit.

Instead of the direct relation from the election, the repository will be
extended with queries for voters. The election can then expose the
results of these queries, or the repository can be directly used.