[!!!][FEATURE] Expose group in voting
Our data model did not contain a direct relation from a voting back to
its group, if a voting belonged to a group. This is not really
feasible in quite a few situations, e.g. when listing the votings of an
election (exposing all grouped votings might not be desired in that
situation, or at least we want to show the relation to the group then).

This commit changes the model back from an MM table to a simple
relation field in the VotingGroup class. The field is only filled if
there is a group, and in that case the election is NULL (because it can
be retrieved via the group). This makes the $votings property of
Election only hold ungrouped votings by default, which is what we want
for that property.

Note that there is no automatic migration for existing data!