[BUGFIX] Some entites are unintendedly removed
When a vote was inserted and removed again (i.e. cancelled before it was
committed), the votes collection within the eligible voter object
triggered a so-called orphan removal in Doctrine. This lead to the
voter and the voting to be removed, which is a very serious issue.

The solution is to make all models aggregate roots by creating
repositories for them. The repositories will not be used internally,
instead we still rely on the existing repositories for our real roots.