[FEATURE] Recording votes of users
This commit implements the Vote class and necessary infrastructure for
it. The votes are created using the VotingService, which in turn uses
stored procedures that create and remove the objects as necessary.
Additionally, transactions are used to make the voting process atomic,
even if more than one vote is being created at the same time.

Error handling for registering the inital vote object is in place, for
committing or cancelling the vote it still has to be implemented.

The VotingService is secured with some functional tests.